+++ src/main/java/sk/upjs/ics/paz1c/fitnesscentrum/dao/ZakaznikDao.java
@@ -9,10 +9,10 @@ public interface ZakaznikDao {
     public List<Zakaznik> dajPritomnychZakaznikov();
     public List<Zakaznik> dajZakaznikovSoZhodouVMene(String vzorka);
     public void pridajZakaznika(Zakaznik zakaznik);
-    public void vymazZakaznika(int idZakaznika);
-    public void prichod(int idZakaznika, Integer idKluca);
-    public void odchod(int idZakaznika);
-    public Zakaznik dajZakaznikaSId(int idZakaznika);
+    public void vymazZakaznika(Long idZakaznika);
+    public void prichod(Long idZakaznika, Long idKluca);
+    public void odchod(Long idZakaznika);
+    public Zakaznik dajZakaznikaSId(Long idZakaznika);
     public Zakaznik dajZakaznikaSCislomPermanentky(String cisloPermanentky);
     
 }
